Make server a standalone binary
